refactor: leave mod.rs empty
